+++ src/Controller/JobsController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Jobs;
 use App\Form\JobsType;
 use App\Repository\JobsRepository;
+use App\Repository\UsersRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -20,21 +21,26 @@ class JobsController extends AbstractController
      */
     public function index(JobsRepository $jobsRepository): Response
     {
+        $user = $this->getUser();
+        $jobs = $jobsRepository->findBy(['createdBy' => $user]);
         return $this->render('jobs/index.html.twig', [
-            'jobs' => $jobsRepository->findAll(),
+            'jobs' => $jobs,
         ]);
     }
 
     /**
      * @Route("/new", name="jobs_new", methods={"GET","POST"})
      */
-    public function new(Request $request): Response
+    public function new(Request $request, UsersRepository $usersRepository): Response
     {
         $job = new Jobs();
         $form = $this->createForm(JobsType::class, $job);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $job->setValidated(false);
+            $job->setPublished(false);
+            $job->setCreatedBy($usersRepository->findOneBy(['email' => $this->getUser()->getUserIdentifier()]));
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($job);
             $entityManager->flush();
@@ -53,9 +59,14 @@ class JobsController extends AbstractController
      */
     public function show(Jobs $job): Response
     {
-        return $this->render('jobs/show.html.twig', [
-            'job' => $job,
-        ]);
+        //check if job was created by the logged in user
+        if ($job->getCreatedBy() == $this->getUser()) {
+            return $this->render('jobs/show.html.twig', [
+                'job' => $job,
+            ]);
+        }else {
+            return $this->redirectToRoute('home');
+        }
     }
 
     /**
@@ -63,19 +74,24 @@ class JobsController extends AbstractController
      */
     public function edit(Request $request, Jobs $job): Response
     {
-        $form = $this->createForm(JobsType::class, $job);
-        $form->handleRequest($request);
+        //check if job was created by the logged in user
+        if ($job->getCreatedBy() == $this->getUser()) {
+            $form = $this->createForm(JobsType::class, $job);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('jobs_index', [], Response::HTTP_SEE_OTHER);
+                return $this->redirectToRoute('jobs_index', [], Response::HTTP_SEE_OTHER);
+            }
+
+            return $this->renderForm('jobs/edit.html.twig', [
+                'job' => $job,
+                'form' => $form,
+            ]);
+        }else {
+            return $this->redirectToRoute('home');
         }
-
-        return $this->renderForm('jobs/edit.html.twig', [
-            'job' => $job,
-            'form' => $form,
-        ]);
     }
 
     /**
@@ -83,12 +99,16 @@ class JobsController extends AbstractController
      */
     public function delete(Request $request, Jobs $job): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$job->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($job);
-            $entityManager->flush();
+        //check if job was created by the logged in user
+        if ($job->getCreatedBy() == $this->getUser()) {
+            if ($this->isCsrfTokenValid('delete' . $job->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($job);
+                $entityManager->flush();
+            }
+            return $this->redirectToRoute('jobs_index', [], Response::HTTP_SEE_OTHER);
+        } else {
+            return $this->redirectToRoute('home');
         }
-
-        return $this->redirectToRoute('jobs_index', [], Response::HTTP_SEE_OTHER);
     }
 }
